requirements.txt
# Core Framework
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# Machine Learning & AI
tensorflow==2.13.0
keras==2.13.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
dlib-bin==19.24.2
imutils==0.5.4
Pillow==10.0.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Video Processing
moviepy==1.0.3
imageio==2.31.1
imageio-ffmpeg==0.4.8

# Utilities
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# File Handling
werkzeug==2.3.7
python-magic==0.4.27
ffmpeg-python==0.2.0

# Development
jupyter==1.0.0
ipython==8.14.0
notebook==7.0.4